backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3

# Cloudinary
cloudinary==1.38.0

# Razorpay
razorpay==1.4.1

# Utils
python-dotenv==1.0.0
httpx==0.26.0
